+++ param/depends.py
@@ -1,4 +1,3 @@
-import inspect
 import weakref
 
 from collections import defaultdict
@@ -204,133 +203,3 @@ def depends(func, *dependencies, watch=False, on_init=False, **kw):
     _depends._dinfo = _dinfo
 
     return _depends
-
-def bind(function, *args, watch=False, **kwargs):
-    """
-    Given a function, returns a wrapper function that binds the values
-    of some or all arguments to Parameter values and expresses Param
-    dependencies on those values, so that the function can be invoked
-    whenever the underlying values change and the output will reflect
-    those updated values.
-
-    As for functools.partial, arguments can also be bound to constants,
-    which allows all of the arguments to be bound, leaving a simple
-    callable object.
-
-    Arguments
-    ---------
-    function: callable
-        The function to bind constant or dynamic args and kwargs to.
-    args: object, param.Parameter
-        Positional arguments to bind to the function.
-    watch: boolean
-        Whether to evaluate the function automatically whenever one of
-        the bound parameters changes.
-    kwargs: object, param.Parameter
-        Keyword arguments to bind to the function.
-
-    Returns
-    -------
-    Returns a new function with the args and kwargs bound to it and
-    annotated with all dependencies.
-    """
-    from .reactive import reactive
-
-    args, kwargs = (
-        tuple(transform_dependency(arg) for arg in args),
-        {key: transform_dependency(arg) for key, arg in kwargs.items()}
-    )
-    dependencies = {}
-
-    # If the wrapped function has a dependency add it
-    fn_dep = transform_dependency(function)
-    if isinstance(fn_dep, Parameter) or hasattr(fn_dep, '_dinfo'):
-        dependencies['__fn'] = fn_dep
-
-    # Extract dependencies from args and kwargs
-    for i, p in enumerate(args):
-        if hasattr(p, '_dinfo'):
-            for j, arg in enumerate(p._dinfo['dependencies']):
-                dependencies[f'__arg{i}_arg{j}'] = arg
-            for kw, kwarg in p._dinfo['kw'].items():
-                dependencies[f'__arg{i}_arg_{kw}'] = kwarg
-        elif isinstance(p, Parameter):
-            dependencies[f'__arg{i}'] = p
-    for kw, v in kwargs.items():
-        if hasattr(v, '_dinfo'):
-            for j, arg in enumerate(v._dinfo['dependencies']):
-                dependencies[f'__kwarg_{kw}_arg{j}'] = arg
-            for pkw, kwarg in v._dinfo['kw'].items():
-                dependencies[f'__kwarg_{kw}_{pkw}'] = kwarg
-        elif isinstance(v, Parameter):
-            dependencies[kw] = v
-
-    def combine_arguments(wargs, wkwargs, asynchronous=False):
-        combined_args = []
-        for arg in args:
-            if hasattr(arg, '_dinfo'):
-                arg = eval_function_with_deps(arg)
-            elif isinstance(arg, Parameter):
-                arg = getattr(arg.owner, arg.name)
-            combined_args.append(arg)
-        combined_args += list(wargs)
-
-        combined_kwargs = {}
-        for kw, arg in kwargs.items():
-            if hasattr(arg, '_dinfo'):
-                arg = eval_function_with_deps(arg)
-            elif isinstance(arg, Parameter):
-                arg = getattr(arg.owner, arg.name)
-            combined_kwargs[kw] = arg
-        for kw, arg in wkwargs.items():
-            if asynchronous:
-                if kw.startswith('__arg'):
-                    combined_args[int(kw[5:])] = arg
-                elif kw.startswith('__kwarg'):
-                    combined_kwargs[kw[8:]] = arg
-                continue
-            elif kw.startswith('__arg') or kw.startswith('__kwarg') or kw.startswith('__fn'):
-                continue
-            combined_kwargs[kw] = arg
-        return combined_args, combined_kwargs
-
-    def eval_fn():
-        if callable(function):
-            fn = function
-        else:
-            p = transform_dependency(function)
-            if isinstance(p, Parameter):
-                fn = getattr(p.owner, p.name)
-            else:
-                fn = eval_function_with_deps(p)
-        return fn
-
-    if inspect.isasyncgenfunction(function):
-        async def wrapped(*wargs, **wkwargs):
-            combined_args, combined_kwargs = combine_arguments(
-                wargs, wkwargs, asynchronous=True
-            )
-            evaled = eval_fn()(*combined_args, **combined_kwargs)
-            async for val in evaled:
-                yield val
-        wrapper_fn = depends(**dependencies, watch=watch)(wrapped)
-        wrapped._dinfo = wrapper_fn._dinfo
-    elif iscoroutinefunction(function):
-        @depends(**dependencies, watch=watch)
-        async def wrapped(*wargs, **wkwargs):
-            combined_args, combined_kwargs = combine_arguments(
-                wargs, wkwargs, asynchronous=True
-            )
-            evaled = eval_fn()(*combined_args, **combined_kwargs)
-            return await evaled
-    else:
-        @depends(**dependencies, watch=watch)
-        def wrapped(*wargs, **wkwargs):
-            combined_args, combined_kwargs = combine_arguments(wargs, wkwargs)
-            return eval_fn()(*combined_args, **combined_kwargs)
-    wrapped.__bound_function__ = function
-    wrapped.reactive = lambda: reactive(wrapped)
-    _reactive_display_objs.add(wrapped)
-    for name, accessor in _display_accessors.items():
-        setattr(wrapped, name, accessor(wrapped))
-    return wrapped
